refactor(client): moved datatable to shared and updated imports
